[FEAT]: Supabase 클라이언트 파일을 services 디렉토리 하위로 변경, supabase oauth 키 연결
